fixed average to be accurate

+++ csharp-prep/Prep4/Program.cs
@@ -26,7 +26,7 @@ class Program
                 }
             }
         }
-        average = sum/numbers.Count;
+        average = (float)sum/numbers.Count;
         Console.WriteLine($"The sum is: {sum}");
         Console.WriteLine($"The average is: {average}");
         Console.WriteLine($"The largest number is: {max}");
